Enable custom result aggregation via strategy
